requirements.txt
django
celery
redis
channels-redis<4.0
django-ninja
django-allauth
django-otp
qrcode
Pillow
whitenoise
django-channels<4.0
daphne<4.0
gunicorn
python-dotenv
django-tailwind
django-browser-reload
uvicorn[standard]
psycopg2-binary
PyJWT
cryptography
celery-progress
gevent
django-celery-beat

